upload data to high performance gpu memory (or slow cpu writable memory)

+++ glfwtest/helpers/Helpers.h
@@ -35,7 +35,7 @@ namespace Helpers
 	static double RoundToTwoDecimals(double n) 
 	{
 		double d = n * 100.0;
-		int i = d + 0.5;
+		int i = (int)(d + 0.5);
 		d = (float)i / 100.0;
 		return d;
 	}
